+++ CHANGELOG.md
@@ -5,12 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- *Undefined offset* PHP Notice in `Factory` when `Message-Id` header is blank.
 
 ## [4.0.0] - 2019-12-05
 ### Added
 - *symfony/mime* added as a required dependency.
 - Dedicated setter and getter methods for primary `Mail` headers: Message-Id,
-  (Origination) Date, Received, In-Reply-To and References. 
+  (Origination) Date, Sender, Received, In-Reply-To and References. 
 ### Fixed
 - Encoding full mailbox header when Display Name contains extended characters.
   Now only encode the affected part of the Display Name, so the address stays
@@ -24,7 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   `AbstractPart::clearHeader($name)`
 - **BC break**: Removed `Mail::formatAddress()`
 - **BC break**: `Mail::addHeader()` no longer allows setting Message-Id,
-  (Origination) Date, Received, In-Reply-To and References.
+  (Origination) Date, Sender, Received, In-Reply-To and References.
   Use the new dedicated setter methods.
 
 ## [3.2.1] - 2019-07-30